Fix ticket 2232
git-svn-id: https://svn.bgdi.admin.ch/mf-chsdi/trunk/chsdi@975 e100092c-b5e6-494f-8bdf-19c1ff2a4001

+++ chsdi/public/GeoAdmin.ux/Permalink/lib/PermalinkProvider.js
@@ -53,7 +53,7 @@ GeoAdmin.PermalinkProvider = Ext.extend(GeoExt.state.PermalinkProvider, {
                 } else if (k === 'zoom') {
                     map_state[k] = parseFloat(params[k]);
                 } else if (k === 'bgOpacity') {
-                    map_state.complementaryLayer.opacity = parseInt(params.bgOpacity);
+                    map_state.complementaryLayer.opacity = parseFloat(params.bgOpacity);
                 } else if (k === 'bgLayer') {
                     if (params.bgLayer === 'pixelmaps-gray') {
                         params.bgLayer =  'ch.swisstopo.pixelkarte-grau';
